+++ src/my-map.ts
@@ -40,6 +40,9 @@ export class MyMap extends LitElement {
   static styles = unsafeCSS(styles);
 
   // configurable component properties
+  @property({ type: String })
+  id = "map";
+
   @property({ type: Number })
   latitude = 51.507351;
 
@@ -127,9 +130,17 @@ export class MyMap extends LitElement {
   @property({ type: Boolean })
   useScaleBarStyle = false;
 
+  // set class property (map doesn't require any reactivity using @state)
+  map?: Map;
+
+  // called when element is created
+  constructor() {
+    super();
+  }
+
   // runs after the initial render
   firstUpdated() {
-    const target = this.shadowRoot?.querySelector("#map") as HTMLElement;
+    const target = this.renderRoot.querySelector(`#${this.id}`) as HTMLElement;
 
     const useVectorTiles =
       !this.disableVectorTiles && Boolean(this.osVectorTilesApiKey);
@@ -171,6 +182,8 @@ export class MyMap extends LitElement {
       }),
     });
 
+    this.map = map;
+
     // make configurable interactions available
     const draw = configureDraw(this.drawPointer);
     const modify = configureModify(this.drawPointer);
@@ -192,11 +205,20 @@ export class MyMap extends LitElement {
 
       if (this.drawMode) {
         drawingSource.clear();
+
+        this.dispatch("geojsonChange", {});
+        this.dispatch(
+          "areaChange",
+          `0 ${this.areaUnit === "m2" ? "m²" : this.areaUnit}`
+        );
+
         map.addInteraction(draw);
         map.addInteraction(snap);
       }
     };
 
+    // this is an internal event listener, so doesn't need to be removed later
+    // ref https://lit.dev/docs/components/lifecycle/#disconnectedcallback
     button.addEventListener("click", handleReset, false);
 
     const element = document.createElement("div");
@@ -208,6 +230,13 @@ export class MyMap extends LitElement {
     if (!this.hideResetControl) {
       map.addControl(ResetControl);
     }
+
+    // Apply aria-labels to OL Controls for accessibility
+    const olControls: NodeListOf<HTMLButtonElement> | undefined =
+      this.renderRoot?.querySelectorAll(".ol-control button");
+    olControls?.forEach((node) =>
+      node.setAttribute("aria-label", node.getAttribute("title") || "")
+    );
 
     // define cursors for dragging/panning and moving
     map.on("pointerdrag", () => {
@@ -263,18 +292,19 @@ export class MyMap extends LitElement {
 
     // draw interactions
     if (this.drawMode) {
-      // check if single polygon feature was provided to load as the initial drawing
+      // make sure drawingSource is cleared upfront, even if drawGeojsonData is provided
+      drawingSource.clear();
+
+      // load an initial polygon into the drawing source if provided, or start from an empty drawing source
       const loadInitialDrawing =
         Object.keys(this.drawGeojsonData.geometry).length > 0;
+
       if (loadInitialDrawing) {
         let feature = new GeoJSON().readFeature(this.drawGeojsonData, {
           featureProjection: "EPSG:3857",
         });
         drawingSource.addFeature(feature);
-        // fit map to extent of intial feature, overriding zoom & lat/lng center
         fitToData(map, drawingSource, this.drawGeojsonDataBuffer);
-      } else {
-        drawingSource.clear();
       }
 
       map.addLayer(drawingLayer);
@@ -292,7 +322,7 @@ export class MyMap extends LitElement {
         const sketches = drawingSource.getFeatures();
 
         if (sketches.length > 0) {
-          const lastSketchGeom = last(sketches).getGeometry();
+          const lastSketchGeom = last(sketches)?.getGeometry();
 
           this.dispatch(
             "geojsonChange",
@@ -314,23 +344,35 @@ export class MyMap extends LitElement {
       });
     }
 
-    // show snapping points when in drawMode, with vector tile basemap enabled, and at zoom > 20
+    // show snapping points when in drawMode, with vector tile basemap enabled, and at qualifying zoom
     if (
       this.drawMode &&
       Boolean(this.osVectorTilesApiKey) &&
       !this.disableVectorTiles
     ) {
-      map.addLayer(pointsLayer);
-      drawingLayer.setZIndex(1001); // display draw vertices on top of snap points
+      // define zoom threshold for showing snaps (not @property yet because computationally expensive!)
+      const snapsZoom: number = 20;
 
+      // display draw vertices on top of snap points
+      map.addLayer(pointsLayer);
+      drawingLayer.setZIndex(1001);
+
+      // extract snap-able points from the basemap, and display them as points on the map if initial render within zoom
+      if (this.zoom < snapsZoom) {
+        pointsSource.clear();
+        const extent = map.getView().calculateExtent(map.getSize());
+        getSnapPointsFromVectorTiles(osVectorTileBaseMap, extent);
+      }
+
+      // continue to fetch & update snaps as map moves
       map.on("moveend", () => {
         const currentZoom: number | undefined = map.getView().getZoom();
-        if (currentZoom && currentZoom < 20) {
+        if (currentZoom && currentZoom < snapsZoom) {
           pointsSource.clear();
           return;
         }
 
-        // extract snap-able points from the basemap, and display them as points on the map
+        // timeout minimizes snap updates mid-pan/drag
         setTimeout(() => {
           pointsSource.clear();
           const extent = map.getView().calculateExtent(map.getSize());
@@ -343,12 +385,13 @@ export class MyMap extends LitElement {
     if (this.showFeaturesAtPoint && Boolean(this.osFeaturesApiKey)) {
       getFeaturesAtPoint(
         fromLonLat([this.longitude, this.latitude]),
-        this.osFeaturesApiKey
+        this.osFeaturesApiKey,
+        false
       );
 
       if (this.clickFeatures) {
         map.on("singleclick", (e) => {
-          getFeaturesAtPoint(e.coordinate, this.osFeaturesApiKey);
+          getFeaturesAtPoint(e.coordinate, this.osFeaturesApiKey, true);
         });
       }
 
@@ -399,7 +442,13 @@ export class MyMap extends LitElement {
   render() {
     return html`<script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
       <link rel="stylesheet" href="https://cdn.skypack.dev/ol@^6.6.1/ol.css" />
-      <div id="map" tabindex="0" />`;
+      <div id="${this.id}" class="map" tabindex="0" />`;
+  }
+
+  // unmount the map
+  disconnectedCallback() {
+    super.disconnectedCallback();
+    this.map?.dispose();
   }
 
   /**